Move things into App.vue, gets stuff *mostly* working
